Integrated with Database & changed to ScrollView (read more)
ISSUE: Need to find a way to fix a footer of sorts, like in
CreateAuditPage, such that the back and exit button are present at all
times.

Included connecting to the database, inheriting from Screen so that the
ViewAuditsPage class can work with the screen manager when it gets to be
time to make it work with the home page. I don't have time to get to it
tonight.

Also made a scrollview to contain the gridlayout. If it had been left as
a gridlayout alone, only a fixed number of items could be shown in the
list. Now, it can update to contain as many as necessary.

Also changed some naming to add clarity in separating the terms
"Audit" and "AuditTemplate".